+++ ndsl/dsl/dace/stree/optimizations/memlet_helpers.py
@@ -3,7 +3,7 @@ from enum import Enum
 import dace.sdfg.analysis.schedule_tree.treenodes as stree
 from dace.memlet import Memlet
 
-from ndsl.logging import ndsl_log
+from ndsl import ndsl_log
 
 
 class AxisIterator(Enum):
