+++ ndsl/internal/hmm.py
@@ -25,7 +25,7 @@ or configuration limitation.
 import ctypes
 import os
 
-from ndsl.logging import ndsl_log
+from ndsl import ndsl_log
 from ndsl.optional_imports import cupy as cp
 from ndsl.optional_imports import numpy_allocator as np_allocator
