+++ tests/monitor/test_zarr_monitor.py
@@ -1,5 +1,4 @@
 import copy
-import logging
 import tempfile
 from datetime import datetime, timedelta
 
@@ -22,8 +21,6 @@ from ndsl.constants import (
 from ndsl.monitor.zarr_monitor import ZarrMonitor, array_chunks, get_calendar
 from ndsl.optional_imports import zarr
 
-
-logger = logging.getLogger("test_zarr_monitor")
 
 # pace's K_DIMS doesn't check the soil dimension
 ALL_K_DIMS = ("k", "k_interface", "k_soil")
